requirements.txt
antiorm==1.2.1
bcrypt==4.0.1
db==0.1.1
db-sqlite3==0.0.1
engineering-notation==0.8.0
Pillow==9.4.0
tk==0.1.0
